requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
django-extensions==3.2.0
## The following requirements were added by pip freeze:
asgiref==3.5.2
factory-boy==3.2.1
Faker==14.1.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
